refactor: 인터페이스가 상위 수준의 모듈에 의존하도록 패키지 이동
