+++ src/main/java/codesquad/springcafe/article/dao/ArticleMemoryRepository.java
@@ -1,6 +1,7 @@
 package codesquad.springcafe.article.dao;
 
 import codesquad.springcafe.article.Article;
+import codesquad.springcafe.article.IArticleRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Repository;
@@ -12,7 +13,7 @@ import java.util.Optional;
 import java.util.concurrent.atomic.AtomicInteger;
 
 @Repository
-public class ArticleMemoryRepository implements IArticleRepository{
+public class ArticleMemoryRepository implements IArticleRepository {
     private final Logger log = LoggerFactory.getLogger(getClass());
     private final Map<Integer, Article> articles = new HashMap<>();
     private final AtomicInteger atomicInteger = new AtomicInteger(0);
